% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valgrind-package.R
\name{a_plus_b}
\alias{a_plus_b}
\title{Sum Two Integers}
\usage{
a_plus_b(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{An integer.}
}
\value{
The sum of the two integers.
}
\description{
Takes two integers and returns their sum.
}
\examples{
a_plus_b(1, 2)
}
