% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valgrind-package.R
\docType{package}
\name{valgrind-package}
\alias{valgrind}
\alias{valgrind-package}
\title{Debug Packages Containing C++ Code with 'valgrind'}
\description{
'valgrind' is a system for debugging and profiling that runs on
Unix-based systems. With 'valgrind' you can detect memory leaks in a variety
of ways. This package provides an R interface to the 'valgrind' tool suite.
It allows you to run 'valgrind' from R, retrieve the output, and process it
in R.
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item  The Valgrind Developers (Valgrind tools)
}

}
